+++ src/hydrogen/core/duct_template.clj
@@ -24,14 +24,14 @@
           :cascading-routes (gen-cascading-routes project-ns ["static/root"
                                                               "api/config"
                                                               "api/example"])}
-   :deps '[[cljs-ajax "0.7.5"]
+   :deps '[[cljs-ajax "0.8.0"]
            [day8.re-frame/http-fx "0.1.6"]
            [duct/compiler.sass "0.2.1"]
            [org.clojure/clojurescript "1.10.339"]
-           [re-frame "0.10.6"]
+           [re-frame "0.10.7"]
            [reagent "0.8.1"]
            [secretary "1.2.3"]
-           [hydrogen/module.core "0.1.6"]]
+           [hydrogen/module.core "0.1.8"]]
    :dev-deps '[[day8.re-frame/re-frame-10x "0.3.7"]]
    :templates {;; Client
                "src/{{dirs}}/client.cljs" (resource "core/cljs/client.cljs")